integration/client: rename package to "client"
