+++ integration/client/client_unix_test.go
@@ -16,7 +16,7 @@
    limitations under the License.
 */
 
-package containerd
+package client
 
 import (
 	"testing"
